:recycle: Add unique key to concept schema
Add unique key attribute - A unique key is value that is composed of the values from one or more
concept attributes. It uniquely identifies concept instances of the same
type.

Change is_identifier - returns True if a concept attr is equal to
CONCEPT|<standard concept>|UNIQUE_KEY

Store the concept attributes needed to compose a unique key for each
concept in concept_schema