:recycle: reshape the stage queue
The pipeline wasn't really made to be reusable, but it's sometimes
treated like it is, and that's dangerous. This migrates pieces of the
pipeline so that configuration arguments all go to init, the logger
gets set up right away, and the stages get constructed directly instead
of being put into (and then pulled back out of) bags of parameters.

It resists treating DataIngestPipeline instances as things to modify and
run again and gets us closer to having a "do my stuff" executive
invocation.

This is part one of converting the DataIngestPipeline class into a
straightforward ingest_data function since you basically just call
run right away _anyway_.