:bug: Fix invalid use of pytest context managers
When using pytest.raises as a context manager, normal context manager
rules apply, and the exception raised must be the final line in the
scope of the context manager. Lines of code after that within the scope
of the context manager will not be executed.

A few valid but verbose uses were also changed for consistency.