:recycle: Break dependency on local ID cache
Relying entirely on the local cache makes it harder to move around and
harder to recover. The target api configuration has been redone to make
it possible to query the server for matching entities first. We can
retool flags to prevent clearing the cache when needed. Target service
plugins now get a new LOADER_VERSION variable that can be set to 2 to
indicate that loading should use the new process.

This should be evaluated with the FHIR api as well before merging.